+++ listeners/actions/index.js
@@ -1,4 +1,8 @@
-import { addCommentCallback, retroOverflowCallback } from "./retro-actions.js";
+import {
+  addCommentCallback,
+  retroOverflowCallback,
+  startRetroHomeCallback,
+} from "./retro-actions.js";
 
 /**
  * Registers all action listeners.
@@ -7,4 +11,5 @@ import { addCommentCallback, retroOverflowCallback } from "./retro-actions.js";
 export const register = (app) => {
   app.action("add_comment", addCommentCallback);
   app.action("retro_overflow", retroOverflowCallback);
+  app.action("start_retro_home", startRetroHomeCallback);
 };